syntax tools: Fix reverting discarding pos changes
This changes `erl_syntax:set_pos/2` for `#wrapper` nodes to also update
the `pos` (really the erl_anno annotation) of the wrapped node. This
means the pos/annotation in `#attr` stay in sync with the actual node.

Together with the change to `erl_syntax:revert/1` this means that
reverting a node will correctly keep any changes to the pos/annotation
for both `#tree` and `#wrapper` nodes.