syntax tools: Fix location info of reverted lists
The hidden `nil` node that is generated takes it's location from the
last element of the list, instead of the list  node itself, because that
generally represents the start of the list.

Similarly, the `cons` nodes are given the location of each element.

However, care must be taken to preserve the other annotations in `pos`,
which is really an `erl_anno` annotation.